+++ scripts/project.py
@@ -131,20 +131,25 @@ JS_TARBALL = {
 }
 
 JS_TSCONFIG = {
-    k: v.parent / "tsconfig.json"
+    k: v.parent / "src/tsconfig.json"
     for k, v in JS_PKG_JSON.items()
-    if (v.parent / "tsconfig.json").exists()
+    if (v.parent / "src/tsconfig.json").exists()
 }
 
 JS_TSSRC = {
     k: sorted(
-        [*(v.parent / "src").rglob("*.ts")] + [*(v.parent / "src").rglob("*.tsx")]
+        [
+            *(v.parent.parent.parent / "src").rglob("*.ts"),
+            *(v.parent.parent / "src").rglob("*.tsx"),
+        ]
     )
     for k, v in JS_TSCONFIG.items()
     if (v.parent / "src").exists()
 }
 
-JS_TSBUILDINFO = {k: v.parent / "tsconfig.tsbuildinfo" for k, v in JS_TSCONFIG.items()}
+JS_TSBUILDINFO = {
+    k: v.parent.parent / ".src.tsbuildinfo" for k, v in JS_TSCONFIG.items()
+}
 
 JS_STYLE = {
     k: sorted((v.parent / "style").glob("*.css"))